rename files and add some love to documentation
