+++ lib/eve_online.rb
@@ -85,8 +85,8 @@ require 'eve_online/esi/character_loyalty_points'
 require 'eve_online/esi/character_wallet'
 require 'eve_online/esi/character_assets'
 require 'eve_online/esi/character_attributes'
-require 'eve_online/esi/industry_character_jobs'
-require 'eve_online/esi/industry_corporation_jobs'
+require 'eve_online/esi/character_industry_job'
+require 'eve_online/esi/corporation_industry_job'
 require 'eve_online/esi/server_status'
 
 require 'eve_online/esi/models/base'
