+++ lib/eve_online/esi/character_industry_job.rb
