+++ lib/eve_online/esi/corporation_industry_job.rb
